Add the ability to add new ControlPoint types to existing groups
